+++ pkg/progressreader/progressreader.go
@@ -14,9 +14,9 @@ type Config struct {
 	In         io.ReadCloser // Stream to read from
 	Out        io.Writer     // Where to send progress bar to
 	Formatter  *streamformatter.StreamFormatter
-	Size       int
-	Current    int
-	LastUpdate int
+	Size       int64
+	Current    int64
+	LastUpdate int64
 	NewLines   bool
 	ID         string
 	Action     string
@@ -29,11 +29,11 @@ func New(newReader Config) *Config {
 
 func (config *Config) Read(p []byte) (n int, err error) {
 	read, err := config.In.Read(p)
-	config.Current += read
-	updateEvery := 1024 * 512 //512kB
+	config.Current += int64(read)
+	updateEvery := int64(1024 * 512) //512kB
 	if config.Size > 0 {
 		// Update progress for every 1% read if 1% < 512kB
-		if increment := int(0.01 * float64(config.Size)); increment < updateEvery {
+		if increment := int64(0.01 * float64(config.Size)); increment < updateEvery {
 			updateEvery = increment
 		}
 	}
